requirements.txt
# Taleemabad Financial Model Dashboard
# Dependencies for portfolio-tracker

# Excel generation
openpyxl>=3.1.0

# Streamlit dashboard
streamlit>=1.29.0

# Data visualization
plotly>=5.18.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# PDF generation
reportlab>=4.0.0

# Google Sheets integration (optional)
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
